Added a helper wrapper class to complement the Label Command
