+++ src/main/java/seedu/address/logic/parser/LabelCommandParser.java
@@ -14,7 +14,7 @@ import seedu.address.logic.parser.exceptions.ParseException;
 /**
  * Parses input arguments and creates a new LabelCommand object
  */
-public class LabelCommandParser {
+public class LabelCommandParser implements Parser<LabelCommand> {
 
     /**
      * Parses the given {@code String} of arguments in the context of the LabelCommand
